+++ libs/contract/src/lib/contract/list/internal-sales/internal-sales.component.ts
@@ -4,9 +4,8 @@ import { filter, map, shareReplay, startWith } from 'rxjs/operators';
 import { UntypedFormControl } from '@angular/forms';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
-import { isInitial } from '@blockframes/contract/negotiation/utils';
 import { capitalize } from '@blockframes/utils/helpers';
-import { Contract, ContractStatus, DetailedContract } from '@blockframes/model';
+import { Contract, ContractStatus, DetailedContract, isInitial } from '@blockframes/model';
 
 @Component({
   selector: 'internal-sales-list',
