+++ libs/contract/src/lib/negotiation/utils.ts
@@ -1,15 +1,5 @@
 import { centralOrgId } from '@env';
-import { Negotiation,  } from '@blockframes/model';
-
-export function isInitial(negotiation: Partial<Negotiation>) {
-  if (!negotiation?._meta) return true;
-  const initial = negotiation.initial;
-  const createdAt = negotiation._meta?.createdAt;
-  initial?.setSeconds(0, 0);
-  createdAt?.setSeconds(0, 0);
-
-  return initial?.getTime() === createdAt?.getTime();
-}
+import { Negotiation } from '@blockframes/model';
 
 export function getReviewer(negotiation: Negotiation) {
   return negotiation.stakeholders.find(
